% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{basic_query}
\alias{basic_query}
\title{Basic Query}
\usage{
basic_query(url, col.spec)
}
\arguments{
\item{url}{The query URL.}
}
\value{
A dataframe.
}
\description{
Helper function for CDEC query handling.
}
\keyword{internal}
