% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.r
\name{cdec-search}
\alias{cdec-search}
\alias{cdec_search_groups}
\alias{cdec_search__stations}
\alias{cdec_map}
\title{CDEC Search Pages
Open the CDEC station/group search page or map interface.}
\usage{
cdec_search_groups()

cdec_search__stations()

cdec_map()
}
\description{
CDEC Search Pages
Open the CDEC station/group search page or map interface.
}
\section{Functions}{
\itemize{
\item \code{cdec_search_groups}: CDEC Group search page.

\item \code{cdec_search__stations}: CDEC Station search page.

\item \code{cdec_map}: CDEC Map locator page.
}}

