% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.r
\name{cdec_data_flags}
\alias{cdec_data_flags}
\title{CDEC Data Flag Information}
\usage{
cdec_data_flags(...)
}
\arguments{
\item{...}{Additional arguments to pass to \code{browseURL}.}
}
\description{
Open the data flag listing page.
}
