% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.r
\name{cdec_meta}
\alias{cdec_meta}
\title{CDEC Station Metadata}
\usage{
cdec_meta(station, ...)
}
\arguments{
\item{station}{The three-letter station code.}

\item{...}{Additional arguments to pass to \code{browseURL}.}
}
\description{
Open the Metadata page for a station.
}
