% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.r
\name{cdec_sensors}
\alias{cdec_sensors}
\title{CDEC Sensor Information}
\usage{
cdec_sensors(...)
}
\arguments{
\item{...}{Additional arguments to pass to \code{browseURL}.}
}
\description{
Open the Sensor listing page.
}
