% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\docType{package}
\name{cder-package}
\alias{cder}
\alias{cder-package}
\title{R Interface to CDEC}
\description{
Provides CDEC data download capabilities from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mkoohafkan/cimir}
  \item Report bugs at \url{https://github.com/mkoohafkan/cimir/issues}
}

}
\author{
\strong{Maintainer}: Michael Koohafkan \email{michael.koohafkan@gmail.com}

}
