% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\name{Deprecated}
\alias{Deprecated}
\alias{cder-deprecated}
\alias{cder_query}
\title{Deprecated functions in package \pkg{cder}.}
\usage{
cder_query(stations, sensors, durations, start.date, end.date)
}
\description{
The functions listed below are deprecated. Alternative
functions with similar
functionality are mentioned. Help pages for deprecated functions are
available at \code{help("cder-deprecated")}.
}
\section{\code{cder_query}}{

For \code{cder_query}, use \code{\link{cdec_query}}.
}

\keyword{internal}
