% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\docType{data}
\name{group.spec}
\alias{group.spec}
\title{Column specification for group data queries.}
\format{
An object of class \code{col_spec} of length 2.
}
\usage{
group.spec
}
\description{
Column specification for group data queries.
}
\keyword{internal}
