% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\docType{data}
\name{station.spec}
\alias{station.spec}
\title{Column specification for station data queries.}
\format{An object of class \code{col_spec} of length 2.}
\usage{
station.spec
}
\description{
Column specification for station data queries.
}
\keyword{internal}
