% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\docType{package}
\name{cder}
\alias{cder}
\alias{cder-package}
\title{Interface to the California Data Exchange Center}
\description{
Access the CDEC Web Service from R.
}
\section{Package options}{


cder uses the following \code{\link[=options]{options()}} to configure behavior:

\itemize{
\item \code{cder.timeout}: The maximum time to wait for a response
from the CDEC Web API.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mkoohafkan/cder}
  \item Report bugs at \url{https://github.com/mkoohafkan/cder/issues}
}

}
\author{
\strong{Maintainer}: Michael Koohafkan \email{michael.koohafkan@gmail.com}

}
